+++ fv3/io/module_wrt_grid_comp.F90
@@ -1720,7 +1720,7 @@
       real(kind=8)  :: wbeg,wend
 
       logical                               :: use_parallel_netcdf
-      real, allocatable                     :: output_fh(:)
+      real(8), allocatable                     :: output_fh(:)
       logical                               :: is_restart_bundle, restart_written
       integer                               :: tileCount
       type(ESMF_Info)                       :: fcstInfo, wrtInfo
